actually let's not cast
commit 267cb8dee1e28fe6d1bc3edc6655ab5d7d7745b6
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Sat Dec 16 04:57:04 2023 -0800

    actually let's not cast

+++ library/contest/random.hpp
@@ -1,13 +1,13 @@
 /** @file */
 #pragma once
-mt19937 rng((unsigned int)chrono::steady_clock::now().time_since_epoch().count());
+mt19937 rng(chrono::steady_clock::now().time_since_epoch().count());
 /**
  * @see https://codeforces.com/blog/entry/61675
  * @code{.cpp}
  *     unsigned int x = rng();
  *
  *     // 64-bit rng
- *     mt19937_64 rng((unsigned int)chrono::steady_clock::now().time_since_epoch().count());
+ *     mt19937_64 rng(chrono::steady_clock::now().time_since_epoch().count());
  *     unsigned long long x = rng();
  *
  *     // fixed seed for debugging
